Introduced a concept of independent steppers that run in parallel to toolhead (and extruder) but without affecting toolhead (like a manual stepper that needs toolhead to stop all movement).
There are several cases that could happen:
If an independent stepper is used when there is no queued move for toolhead or extruder, then the independent stepper starts moving immediately.
If an independent stepper is used, and there is a queued move for toolhead or extruder, then the independent stepper starts moving right after dequeue move that is currently at the end of the toolhead queue,
If an independent stepper is used while this independent stepper is currently moving, then this new move will be planned right after it.