Added waiting with disabling independent steppers until all steps are transmitted.
The standard way how stepper motors are enabled or disabled is that the message for disabling is transmitted immediately into the MCU.
But steps for stepper motors are firstly compressed and then added into the message queue for a particular stepper.
So, it could happen that there are some pending steps or messages in the queue, and before they are transmitted, the message for disabling stepper motors arrives. That leads to disabling the stepper motor before it starts moving.
Also, messages for the stepper motor need to be transmitted ordered by time. When this order isn't met (for example, by transmitting a message for disabling a stepper at the "wrong" time), then the "Timer too close" issue appears.

So now we disable independent steppers only if there are no pending steps or messages in the queue. This also allows setting delay, after which the independent stepper will be disabled.